+++ scripts/deploySigV2Validator.ts
@@ -5,7 +5,7 @@ const pathOutputJson = path.join(__dirname, './deploy_validator_output.json');
 
 async function main() {
   // const stateAddress = 'TODO:'; // current state smart contract on opt
-  const stateAddress = 'TODO:'; // current state smart contract on opt-sepolia
+  const stateAddress = '0x9a1A258702050BcFB938Ad8Ec0996503473216d1'; // current state smart contract on opt-sepolia
 
   const verifierContractWrapperName = 'VerifierSigWrapper';
   const validatorContractName = 'CredentialAtomicQuerySigV2Validator';
